+++ src/main/java/site/geni/FarLands/mixins/client/block/RepeaterBlockMixin.java
@@ -45,9 +45,9 @@ public abstract class RepeaterBlockMixin extends HorizontalFacingBlockMixin {
 	private void addParticlesProperly(BlockState blockState, World world, BlockPos blockPos, Random random, CallbackInfo ci) {
 		if (Config.getConfig().fixParticles) {
 			Direction direction = blockState.get(field_11177);
-			final double x = (blockPos.getX() + 0.5D) + (random.nextDouble() - 0.5D) * 0.2D;
-			final double y = (blockPos.getY() + 0.4D) + (random.nextDouble() - 0.5D) * 0.2D;
-			final double z = (blockPos.getZ() + 0.5D) + (random.nextDouble() - 0.5D) * 0.2D;
+			final double x = (blockPos.getX() + 0.5D) + ((double) random.nextFloat() - 0.5D) * 0.2D;
+			final double y = (blockPos.getY() + 0.4D) + ((double) random.nextFloat() - 0.5D) * 0.2D;
+			final double z = (blockPos.getZ() + 0.5D) + ((double) random.nextFloat() - 0.5D) * 0.2D;
 
 			float float_1 = -5.0F;
 			if (random.nextBoolean()) {
